Correction bug Anapo plot NA VMS
Correction of a bug in Anapo plot creation concerning NA in VMS data.
VMS information (position, date, time, ...) all be missing if it comes from the study of an activity at the port. In this case, the cross-reference with VMS is not performed in the Anapo control to save time, as the activity is necessarily valid.
Resolution: delete VMS if position is missing in plot